+++ extlib/CGUITTFont.h
@@ -1,9 +1,9 @@
 #ifndef __C_GUI_TTFONT_H_INCLUDED__
 #define __C_GUI_TTFONT_H_INCLUDED__
 
-#include <irrlicht/irrlicht.h>
 #include <ft2build.h>
 #include FT_FREETYPE_H
+#include <irrlicht/irrlicht.h>
 
 namespace irr
 {